feat!: adding themable type classes
BREAKING CHANGE:

This commit adds themable type classes to the component.

It also promoted the 'sm' variant to the default visual style. This effectively means the default 'lg' variant is no longer available, and the component will now always render in the 'sm' style.

As a result, the 'sm' attribute is now deprecated and will be removed in a future version.

This commit also corrects the 'ariaHidden' prop type from 'String' to 'Boolean', as it was being used as a Boolean in the test file & docs but incorrectly labeled.